--- Drawing-POC/package.json
@@ -1,3 +1,0 @@
-{
-    "main": "Drawing-POC.js" 
-}
