+++ Drawing/Drawing.module.less
@@ -1,0 +1,3 @@
+.drawing {
+	/* Defined to prevent global export */
+}
